# docs/openapi/backend.json
{
  "components": {
    "schemas": {
...
      "DataResponse_IdRef_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IdRef"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[IdRef]",
        "type": "object"
      },
...
      "DataResponse_Rule_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Rule"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[Rule]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PagedListResponse_Rule_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/Rule"
            },
            "title": "Data",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "count",
          "total"
        ],
        "title": "PagedListResponse[Rule]",
        "type": "object"
      },
...
      "RuleCreate": {
        "description": "A rule as the create dialog sends it.\n\n``search_term``, ``text_match_option`` and ``filters`` are the\n``/search/global-search`` request being saved, spelled the way that route\ntakes it.\n\n``filters.objects`` is the search's tab rather than a switch anybody set:\nthe dialog opens over a result list, and saving from *Tables* means the rule\nlabels tables. Omitted for the *All* tab, and stored omitted -- see\n:func:`create_rule` -- because \"no kind was picked\" and \"the kinds picked\nwere none\" are not the same rule.",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/GlobalSearchFilters"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "search_term": {
            "title": "Search Term",
            "type": "string"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/RuleTagRef"
            },
            "title": "Tags",
            "type": "array"
          },
          "text_match_option": {
            "default": "contains",
            "title": "Text Match Option",
            "type": "string"
          }
        },
        "required": [
          "name",
          "search_term",
          "tags"
        ],
        "title": "RuleCreate",
        "type": "object"
      },
...
      "RuleUpdate": {
        "description": "The part of a rule an edit may change: its name.\n\nNot the search, and not the tags either. Both of those are what a rule *is*\n-- change one and the rule labels a different set of objects, while the\nlabels it already wrote stay where they are, explaining nothing and\nre-applied by nobody. That is a new rule and a deleted one, which is two\noperations this API already has.\n\nA name, by contrast, is only how the settings list refers to the rule: the\nlabels it wrote name it by id, so renaming reaches all of them without a\nsingle ``tag_target`` row being written.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RuleUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Auto Ontology API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rules": {
      "get": {
        "description": "Return rules, with the tags each applies.\n\n*q*, when given, keeps the rules whose name contains it, and those applying\na tag whose name does -- the two things the settings list shows, so a search\nthat hides a row can be explained by what is on screen.\n\nRules come back ordered by name, case-insensitively, and *skip*/*limit*\nselect one page of that order; ``total`` counts every match so a caller\nknows when to stop asking. Omitting *limit* returns every matching rule.\n\nEmpty for a deployment where nobody has saved one, which the Rules settings\npage shows as its ordinary empty state.",
        "operationId": "list_rules_api_rules_get",
        "parameters": [
          {
            "description": "Case-insensitive substring filter on the rule name or on the name of a tag it applies.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive substring filter on the rule name or on the name of a tag it applies.",
              "title": "Q"
            }
          },
          {
            "description": "Rows to skip before the page starts.",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip before the page starts.",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "Page size, at most 100. Omit to return every matching row in one response.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page size, at most 100. Omit to return every matching row in one response.",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedListResponse_Rule_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Rules",
        "tags": [
          "rules"
        ]
      },
      "post": {
        "description": "Save a rule, and answer with its id.\n\nThe id alone: the dialog that posted the rule closes on success and the\nsettings list re-reads, so a whole rule in the answer would be a copy\nnothing renders. Everything else about a stored rule -- the timestamps, the\nauthor, the tag names -- comes from :func:`list_rules`, read from the row\nrather than from what was posted.\n\nThe filters are stored as the caller sent them, ``exclude_none`` and not a\nfull dump: a filter that was not sent is absent from the stored JSON rather\nthan present as ``null``. It reads back the same either way -- the response\nmodel fills the default -- but the row then says which filters a rule\nactually carries, instead of recording every filter the search has ever\noffered against every rule ever saved.\n\nSaving the rule also *applies* it: the search is replayed and the tags are\nattached to everything it matches, recorded as this rule's doing so a tag's\npage can name the rule that labelled each object. The insert and those\nlabels are one transaction, so a rule that could not label the catalog is\nnot left saved and inert -- the create either produces a rule with its\nlabels or nothing at all. The search that finds them runs before that\ntransaction, holding no write connection and no locks while it does; see\n``rules.service.find_targets``.\n\n400 for a body that could never be a rule: a blank name, a search term\nshorter than global search accepts, a match option it does not implement,\nor no tags at all. 404 for a tag id that is not a tag. 409 for a name\nanother rule already holds, as renaming answers. 401 when the request\ncarries no identity to attribute the rule to.",
        "operationId": "create_rule_api_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IdRef_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Rule",
        "tags": [
          "rules"
        ]
      }
    },
    "/api/rules/{rule_id}": {
      "delete": {
        "description": "Delete one rule by id.\n\n**This also takes back every tag the rule applied**, unless ``keep_tags``\nsays otherwise. The labels it wrote cascade with it -- see\n``rules_dal.delete_rule`` -- while labels people applied by hand stay,\nincluding on an object the rule had also matched, because those rows record\nno rule to cascade from. A rule goes on labelling the catalog as the catalog\ngrows, so by default its labels are not left behind for nothing on screen to\nexplain and nothing to re-apply.\n\n**And a tag left labelling nothing is deleted**, unless another rule applies\nit: emptying it is what this request just did, and a tag that describes no\npart of the catalog and that nothing will apply again is not vocabulary. One\nthat still labels anything -- another object, or the same one by a person's\nhand -- stays.\n\n``keep_tags`` is the other reading, and a defensible one: the labels are a\nclassification that happens to have come from a rule, worth keeping once the\nrule that suggested them is gone. They stay where they are, attributed to\nwhoever wrote the rule -- a rule is a person deciding a search's results\ndeserve a tag, so their own labels is what is left when the mechanism goes,\nand the tag's page names them there. Not the caller, who may be an admin\ndeleting somebody else's rule. Recreating the same rule afterwards does not\nadopt them -- an existing label is left alone rather than claimed, the same\nway a rule cannot take over a hand-applied one -- so this is a way to keep\nthe labels, not a way to detach and reattach them.\n\n404 rather than a silent 204 for an id that is not there: the settings page\ndeletes from a list it has already read, so a missing rule means that list\nis stale and the row would be left on screen with nothing to explain it.\n\nAnswers with the id, as deleting a tag does, so a caller has the row it just\nremoved without holding on to what it sent.",
        "operationId": "delete_rule_api_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          },
          {
            "description": "Leave the tags the rule applied on their objects, attributed to whoever wrote the rule, instead of removing them with the rule.",
            "in": "query",
            "name": "keep_tags",
            "required": false,
            "schema": {
              "default": false,
              "description": "Leave the tags the rule applied on their objects, attributed to whoever wrote the rule, instead of removing them with the rule.",
              "title": "Keep Tags",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IdRef_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Rule",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "description": "Rename a rule.\n\nThe name is all an edit changes -- see :class:`RuleUpdate` for why the\nsearch and the tags are not editable. What the rule has already labelled is\nuntouched: those rows name the rule by id, so the new name appears against\nevery object it labelled under the old one.\n\nAnswers with the whole rule rather than an echo of the name, the way\nrenaming a tag does: the edit also advances ``modified`` and records a\n``modified_by``, so the row the caller just edited is redrawn from this one\nresponse instead of from a re-read.\n\n400 for a blank name. 409 for a name another rule holds -- rule names are\nunique, so that the list's own way of referring to a rule identifies one.\nA rename to the rule's own name is not a conflict, and neither is one that\nonly changes case, which is how a rule created shouting gets fixed. 404 for\nan id that is not stored -- the settings page edits a row from a list it has\nalready read, so a missing rule means that list is stale.\n\nNo 401, unlike the create: an edit is recorded against whoever made it when\nthe request says, and a request that names nobody still renames the rule\nsomebody else created. ``modified_by`` is nullable for exactly that.",
        "operationId": "update_rule_api_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Rule_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Rule",
        "tags": [
          "rules"
        ]
      }
    },
...
  }
}